Refactor day 10
No need to track visited nodes in either part because the
strictly-increasing trails prevent backtracking anyway. We also don't
need to care about full trail, only the endpoints - ultimately, the
only difference between the two parts is that Part 1 needs the number of
distinct (start,end) pairs, while Part 2 needs them counted up to
multiplicity - all of which is a job for Counter.